# Optimization for the performance of addModules(), addExplodedModules(), and addFlattenModules()