## Summary
Reduce the number of times to traverse buildpack bits. Save about 27% time spent on `addExplodedModules()`. 
Call `NToLayerTar` in goroutine. Save more than half time spent on `addFlattenModules()`.

## Description

#### Before
Each buildpack bits will be traversed three times: download, calculate diffID and stream to daemon.
`NToLayerTar` in `addFlattenModules()` is called for each `flattenModule` sequentially.

#### After
The diffID of buildpack will calculated while it is downloaded and saved as a tarfile.
`NToLayerTar` is called in goroutine, referring to addExplodedModules().

## Profiling Result

| command | average time spent on addExplodedModules() before | average time spent on addExplodedModules() after | ratio |
| ----------- | ----------------------------- | --------------------------- | ------------- |
| pack build demo-app --builder cnbs/sample-builder:bionic -p ../samples/java/maven --buildpack paketobuildpacks/java:latest | 3.804439914s | 2.746395149s | 0.72189115 |

| command | average time spent on addFlattenModules() before | average time spent on addFlattenModules() after | ratio |
| ----------- | ----------------------------- | --------------------------- | ------------ |
| pack builder create cnbs/sample-builder:alpine --config builder.toml --flatten | 7.761502674s | 3.751928196s | 0.48340229 |

```
# builder.toml
# Buildpacks to include in builder
[[buildpacks]]
uri = "../samples/buildpacks/hello-processes"

[[buildpacks]]
# Packaged buildpacks to include in builder;
# the "hello-universe" package contains the "hello-world" and "hello-moon" buildpacks
uri = "docker://paketobuildpacks/java:latest"

# Order used for detection
[[order]]
    # This buildpack will create a process type "sys-info" to display runtime information
    [[order.group]]
    id = "samples/hello-processes"
    version = "0.0.1"

# Stack that will be used by the builder
[stack]
id = "io.buildpacks.samples.stacks.bionic"
# This image is used at runtime
run-image = "cnbs/sample-stack-run:bionic"
# This image is used at build-time
build-image = "cnbs/sample-stack-build:bionic"
```

## Documentation
<!-- If this change should be documented, please create an issue or PR on https://github.com/buildpacks/docs and link below. -->
<!-- NOTE: This can be added (by editing the issue) after the PR is opened. -->

- Should this change be documented?
    - [ ] Yes, see #___
    - [ ] No

## Related
<!-- If this PR addresses an issue, please provide issue number below. -->

Resolves #___