@natalieparellano 2023-05-19
As per usual it seems Codecov is being a bit too picky here... I wonder if there is a way we can get it to ignore error checks, at least on i/o operations 🤔