@jjbustamante 2023-08-30
I am closing this PR based on the latest conversation regarding this feature and the after the [RFC](https://github.com/buildpacks/rfcs/pull/290) is approved, we will [remove](https://github.com/buildpacks/rfcs/pull/290#issuecomment-1660543031) most the current implementation and this is not going to be valid anymore